requirements.txt
psycopg2
flask
gunicorn
pyyaml